// TimeWizard_fe/src/app/Core/models/formated-data.interface.ts
export interface FormattedData {
    inputs: InputData[];
}

export interface InputData {
    s_frontend: number;
    s_backend: number;
    lf_react: number;
    lf_angular: number;
    lf_javascript: number;
    lb_node: number;
    lb_dotnet: number;
    lb_java: number;
    Complexity_value: number;
}